malAtomProperty(): BATatoms{}.storage is of type int, not of type long;
actually, I am not sure wether this code/assignemt makes any sense at all, here;
while all other BATatoms[] members assigned here are function pointers,
BATatoms[].storage is an integer ... ?

+++ monetdb5/mal/mal_atom.c
@@ -121,7 +121,7 @@ int malAtomProperty(MalBlkPtr mb, InstrPtr pci)
 		break;
 	case 's':
 		if (idcmp("storage", name) == 0 && pci->argc == 1) {
-			BATatoms[tpe].storage = (*(long (*)(void))pci->fcn)();
+			BATatoms[tpe].storage = (*(int (*)(void))pci->fcn)();
 			setAtomName(pci);
 			return 1;
 		}
